refactor: move figures to common reflectometry module
