+++ src/ess/reflectometry/types.py
@@ -1,5 +1,5 @@
 # Copyright (c) 2025 Scipp contributors (https://github.com/scipp)
-from typing import NewType, TypeVar
+from typing import Any, NewType, TypeVar
 
 import sciline
 import scipp as sc
@@ -70,6 +70,11 @@ WavelengthBins = NewType("WavelengthBins", sc.Variable)
 """Bins for the wavelength histogram, also used to filter the event data."""
 
 
+class ThetaBins(sciline.Scope[RunType, sc.Variable], sc.Variable):
+    """Binning in theta that takes into consideration that some
+    detector pixels have the same theta value."""
+
+
 class Filename(sciline.Scope[RunType, str], str):
     """Filename of an event data nexus file."""
 
@@ -92,8 +97,7 @@ class DetectorSpatialResolution(sciline.Scope[RunType, sc.Variable], sc.Variable
 
 
 class SampleSize(sciline.Scope[RunType, sc.Variable], sc.Variable):
-    # TODO is this radius or total length?
-    """Size of the sample. If None it is assumed to be the same as the reference."""
+    """Diameter of the sample. If None it is assumed to be the same as the reference."""
 
 
 class ProtonCurrent(sciline.Scope[RunType, sc.DataArray], sc.DataArray):
@@ -114,3 +118,9 @@ BeamDivergenceLimits = NewType("BeamDivergenceLimits", tuple[sc.Variable, sc.Var
 
 ReferenceFilePath = NewType("ReferenceFilePath", str)
 """Path to the cached normalization matrix"""
+
+
+WavelengthThetaFigure = NewType("WavelengthThetaFigure", Any)
+WavelengthZIndexFigure = NewType("WavelengthZIndexFigure", Any)
+QThetaFigure = NewType("QThetaFigure", Any)
+ReflectivityDiagnosticsView = NewType("ReflectivityDiagnosticsView", Any)
